Move `proto::message::Unparsed` to own module
